dz4/src/requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
pandas
python-geohash